fix: 모델 모듈화
