+++ util/labels.py
@@ -1,0 +1,7 @@
+LABEL2IDX = {
+    '감사합니다': 0, '괜찮다': 1, '내일': 2, '너': 3, '네': 4, '누가': 5,
+    '만들다': 6, '만족': 7, '말하다': 8, '뭐': 9, '시작': 10, '아니요': 11,
+    '안녕하세요': 12, '어디': 13, '어떻게': 14, '언제': 15, '오늘': 16,
+    '왜': 17, '의견': 18, '이해': 19, '자리': 20, '잠시': 21, '저': 22,
+    '좋다': 23, '죄송합니다': 24, '지금': 25, '질문': 26, '학교': 27, '회의': 28
+}
